requirements.txt
# ─────────────────────────────────────────────────────────────
# S-Core Backend Dependencies
# ─────────────────────────────────────────────────────────────

# FastAPI & Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# AI & LLM
groq>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector Database
chromadb>=0.4.0

# MongoDB
pymongo>=4.6.0

# Internet Search
duckduckgo-search>=4.0.0

# Document Processing - PDF
pdfplumber>=0.10.0

# Document Processing - Word
python-docx>=1.0.0

# Document Processing - PowerPoint
python-pptx>=0.6.21

# Image Processing
pytesseract>=0.3.10
pillow>=10.0.0

# Utilities
tqdm>=4.66.0
